backend/requirements.txt
fastapi 
uvicorn
sqlalchemy
psycopg2
alembic
pydantic
pydantic_settings
python-jose
passlib[crypt]